Rename the rdata modules to use names instead of numbers.
